Struct ffmpeg::software::resampling::context::Context

pub struct Context { /* fields omitted */ }

Methods

impl Context

fn get(src_format: Sample,
src_channel_layout: ChannelLayout,
src_rate: u32,
dst_format: Sample,
dst_channel_layout: ChannelLayout,
dst_rate: u32)
-> Result<Self, Error>

Create a resampler with the given definitions.

fn input(&self) -> &Definition

Get the input definition.

fn output(&self) -> &Definition

Get the output definition.

fn delay(&self) -> Option<Delay>

Get the remaining delay.

fn run(&mut self,
input: &Audio,
output: &mut Audio)
-> Result<Option<Delay>, Error>

Run the resampler from the given input to the given output.

When there are internal frames to process it will return Ok(Some(Delay { .. })).

fn flush(&mut self, output: &mut Audio) -> Result<Option<Delay>, Error>

Convert one of the remaining internal frames.

When there are no more internal frames Ok(None) will be returned.